Reorganize high level API
