+++ src/pyobo/__init__.py
@@ -2,13 +2,15 @@
 
 """A python package for handling and generating OBO."""
 
-from .config import get_config  # noqa: F401
-from .extract import (  # noqa: F401
-    get_alts_to_id, get_ancestors, get_descendants, get_filtered_properties_mapping, get_filtered_relations_df,
-    get_filtered_xrefs, get_hierarchy, get_id_name_mapping, get_id_species_mapping, get_id_synonyms_mapping, get_name,
-    get_name_by_curie, get_name_id_mapping, get_primary_curie, get_primary_identifier, get_property, get_relations_df,
-    get_species, get_subhierarchy, get_xref, get_xrefs_df,
+from .api import (  # noqa: F401
+    get_alts_to_id, get_ancestors, get_descendants, get_filtered_properties_df,
+    get_filtered_properties_mapping, get_filtered_properties_multimapping, get_filtered_relations_df,
+    get_filtered_xrefs, get_hierarchy, get_id_name_mapping, get_id_species_mapping, get_id_synonyms_mapping,
+    get_id_to_alts, get_name, get_name_by_curie, get_name_id_mapping, get_primary_curie, get_primary_identifier,
+    get_properties, get_properties_df, get_property, get_relations_df, get_species, get_subhierarchy,
+    get_typedef_id_name_mapping, get_xref, get_xrefs_df,
 )
+from .config import get_config  # noqa: F401
 from .getters import get  # noqa: F401
 from .identifier_utils import normalize_curie, normalize_prefix  # noqa: F401
 from .normalizer import OboNormalizer, ground  # noqa: F401
