+++ src/pyobo/cli/lookup.py
@@ -8,10 +8,10 @@ import click
 from more_click import verbose_option
 
 from .utils import echo_df, force_option, prefix_argument
-from ..extract import (
-    get_ancestors, get_descendants, get_filtered_properties_df, get_filtered_relations_df, get_filtered_xrefs,
-    get_hierarchy, get_id_name_mapping, get_id_synonyms_mapping, get_id_to_alts, get_name, get_name_by_curie,
-    get_properties_df, get_relations_df, get_typedef_id_name_mapping, get_xrefs_df,
+from ..api import (
+    get_ancestors, get_descendants, get_filtered_properties_df, get_filtered_relations_df,
+    get_filtered_xrefs, get_hierarchy, get_id_name_mapping, get_id_synonyms_mapping, get_id_to_alts, get_name,
+    get_name_by_curie, get_properties_df, get_relations_df, get_typedef_id_name_mapping, get_xrefs_df,
 )
 from ..identifier_utils import normalize_curie
