+++ src/pyobo/sources/pid.py
@@ -9,7 +9,7 @@ from typing import Iterable, List, Mapping, Tuple
 import pandas as pd
 from protmapper.uniprot_client import get_gene_name, get_hgnc_id
 
-from ..extract import get_id_name_mapping
+from ..api import get_id_name_mapping
 from ..ndex_utils import CX, ensure_ndex_network_set, iterate_aspect
 from ..struct import Obo, Reference, Term
 from ..struct.typedef import has_part
