+++ src/pyobo/sources/reactome.py
@@ -9,8 +9,8 @@ import bioversions
 import pandas as pd
 from tqdm import tqdm
 
+from ..api import get_name_id_mapping
 from ..constants import SPECIES_REMAPPING
-from ..extract import get_name_id_mapping
 from ..io_utils import multidict
 from ..path_utils import ensure_df
 from ..struct import Obo, Reference, Term, from_species, has_part
