+++ tox.ini
@@ -62,7 +62,7 @@ deps =
 commands =
     flake8 \
         src/pyobo/getters.py \
-        src/pyobo/extract.py
+        src/pyobo/api/
 description = Run the flake8 darglint tool.
 
 [testenv:pyroma]
@@ -107,6 +107,10 @@ commands =
     sphinx-build -W -b coverage -d {envtmpdir}/build/doctrees {envtmpdir}/source {envtmpdir}/build/coverage
     cat {envtmpdir}/build/coverage/c.txt
     cat {envtmpdir}/build/coverage/python.txt
+whitelist_externals =
+    /bin/cp
+    /bin/cat
+    /bin/mkdir
 
 [testenv:coverage-report]
 deps = coverage
